requirements.txt
asyncpraw==7.1.1
asyncio==3.4.3
discord.py==1.7.1
Pillow==8.2.0
sympy==1.6.2
matplotlib==3.3.1
chempy==0.7.12
chess==1.3.0
pymongo==3.11.2
dnspython==2.0.0
motor==2.3.0
psutil==5.7.3
dblpy==0.4.0
gTTS==2.2.2
python-dotenv==0.14.0
pilmoji
emoji